chore(passport): Revise "continue with" login experience
